% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Significant Earthquake Data.}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Dataframe containing earthquake data}
}
\value{
Cleaned dataframe
}
\description{
Clean NOAA Significant Earthquake Data.
}
\examples{
# Assumes the NOAA earthquake dataset has been obtained per method
# in `data-raw/earthquakes.R`.
eq_clean_data(earthquakes)

}
