% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Plots earthquake events on a map.
Helper function to create html labels for events.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Earthquake dataframe}
}
\value{
List of html labels for events
}
\description{
Plots earthquake events on a map.
Helper function to create html labels for events.
}
\examples{
\dontrun{
# Assumes data has been downloaded as `earthquakes.tsv`.
readr::read_delim("./earthquakes.tsv", delim = "\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::mutate(popup_text = eq_create_label(.))
}

}
