% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Makes the NOAA Significant Earthquake data easier to use.}
\usage{
eq_location_clean(loc)
}
\arguments{
\item{loc}{Earthquake location name}
}
\value{
Cleaned location name
}
\description{
\itemize{
\item Combines date columns into a single \code{Date} column.
\item Extracts countries from the location column.
\item If location is a US state, sets country to USA.
Clean location names.
}
}
\details{
Removes country, etc. from locations.
}
\examples{
eq_location_clean("CALIFORNIA:  SANTA BARBARA")
eq_location_clean("MEXICO:  BAJA CALIFORNIA: LORETO")

}
