% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Map of earthquake events.}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Earthquake dataframe}

\item{annot_col}{Name of column containing data for event popups}
}
\value{
Leaflet map
}
\description{
Map of earthquake events.
}
\examples{
# Assumes the NOAA earthquake dataset has been obtained per method
# in `data-raw/earthquakes.R`.
eq_map(
  dplyr::mutate(
    eq_clean_data(earthquakes),
    popup_text = eq_create_label(.data)
  ),
  annot_col = "popup_text"
)

}
