% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Plot labels for NOAA Significant Earthquake Data on a timeline.}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = NA,
  ...
)
}
\description{
Plot labels for NOAA Significant Earthquake Data on a timeline.
}
\examples{
\dontrun{
# Assumes data has been downloaded as `earthquakes.tsv`.
readr::read_delim("./earthquakes.tsv", delim = "\t") \%>\%
  eq_clean_data(df) \%>\%
  ggplot(aes(x = Date, y = Country, colour = Deaths),  alpha = 1) +
  geom_timeline() +
  geom_timeline_label(
    aes(label = `Location Name`, magnitude = Mag),
    n_max = 5
  )
}

}
