% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Custom geom for plotting NOAA Significant Earthquake data on a timeline.}
\description{
Custom geom for plotting NOAA Significant Earthquake data on a timeline.
}
\keyword{datasets}
