% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Custom geom for plotting NOAA Significant Earthquake data labels on
a timeline.}
\description{
Custom geom for plotting NOAA Significant Earthquake data labels on
a timeline.
}
\keyword{datasets}
