% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Significant Earthquake Data.}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Dataframe containing earthquake data}
}
\value{
Cleaned dataframe
}
\description{
Clean NOAA Significant Earthquake Data.
}
\examples{
\dontrun{
# Assumes data has been downloaded as `earthquakes.tsv`.
readr::read_delim("./earthquakes.tsv", delim = "\t") \%>\%
  eq_clean_data(df)
}

}
