% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Plots earthquake events on a map.
Helper function to create html labels for events.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Earthquake dataframe}
}
\value{
List of html labels for events
}
\description{
Plots earthquake events on a map.
Helper function to create html labels for events.
}
\examples{
# Assumes the NOAA earthquake dataset has been obtained per method
# in `data-raw/earthquakes.R`.
dplyr::mutate(
  eq_clean_data(earthquakes),
  popup_text = eq_create_label(.data)
)

}
